+++ ui/fx/src/main/kotlin/org/pcsoft/framework/simplay/fx/PaperSheetView.kt
@@ -536,6 +536,8 @@ class PaperSheetView : Control() {
         fun moveToPrevBlock()
         fun moveToNextSymbol()
         fun moveToPrevSymbol()
+        fun moveToNextPage()
+        fun moveToPrevPage()
     }
 
     private var caretCommands: CaretCommands? = null